Refactor plugin; add detection bus & safety utils
Large refactor and feature additions: reorganized AntiLitematicaPlugin lifecycle (shutdownAll, lang save, admin GUI, metrics), added DetectionBus/DetectionHandler, Bedrock player detection, CommandSanitizer, and unified PlacementGuard per-player trackers. Improved command handling (tab completion, safer player lookup), Protocol/NBT extraction shared via NbtLite, synchronized DetectionLogger, Discord proxy auth, OneBot escaping, and safer ban-hook commands (AbstractBanHook). Improved PunishmentTracker DB migrations and retention config. Docs/config updates (graduated_punishment now disabled by default, README/wiki/commands). Miscellaneous cleanup and safety fixes.

+++ src/main/java/top/chenray/antilitematica/util/TokenBucket.java
@@ -8,7 +8,7 @@ public final class TokenBucket {
 
    private TokenBucket(double refillPerSecond, double capacity) {
       this.refillPerMillis = refillPerSecond / (double)1000.0F;
-      this.capacity = Math.max((double)1.0F, capacity);
+      this.capacity = Math.max(1.0D, capacity);
       this.tokens = this.capacity;
       this.lastRefillMs = System.currentTimeMillis();
    }
